Segregated code out
